refactor: move Sanitizes_Css_Value into Traits namespace; bump variant editor hook to priority 10
